🚧 5FPZAB task: apply external agent result
