+++ .agentplane/tasks/202609021331-5FPZAB/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202609021331-5FPZAB/repair-lifecycle-projection-integrity-after-m3-c",
   "created_at": "2026-09-02T15:35:26.605Z",
-  "diffstat_sha256": "sha256:c2434239a2f7a8bd6b08e364e29c3965ec856ebb3fabb6dd90b60140cdb70846",
+  "diffstat_sha256": "sha256:260cfa2ef3e303dd9718d3196875d05febd9a089683059197ff11327022e1db5",
   "last_verified_at": null,
   "provider": {
     "hostname": "github.com",
