+++ .agentplane/tasks/202609021331-5FPZAB/pr/review.md
@@ -29,37 +29,39 @@ Created: 2026-09-02T15:35:26.605Z
 - Head: computed live by `agentplane pr check` / `agentplane integrate`
 
 ```text
- .../task-backend/task-centric-backend-adapter.ts   | 234 +++++++++++++++------
- .../src/cli/run-cli.core.route-decision.test.ts    |  56 ++++-
- .../branch/cleanup-merged.targeted.test.ts         | 113 +++++++++-
+ .../task-backend/task-centric-backend-adapter.ts   |  72 +---
+ .../task-centric-backend-projection.ts             | 185 ++++++++
+ .../src/cli/run-cli.core.route-decision.test.ts    |  56 ++-
+ .../branch/cleanup-merged.targeted.support.test.ts | 462 ++++++++++++++++++++
+ .../branch/cleanup-merged.targeted.test.ts         | 475 ++++-----------------
  .../src/commands/branch/cleanup-merged.ts          |  15 +-
  .../commands/shared/route-decision-workspace.ts    |   2 +-
- .../shared/task-backend-branch-snapshot.ts         | 123 ++++++++++-
- .../task-backend-branch-snapshot.unit.test.ts      | 136 +++++++++++-
- .../src/commands/shared/task-backend.test.ts       |  33 +--
- .../agentplane/src/commands/shared/task-backend.ts |  55 +++--
- .../src/commands/shared/task-mutation.test.ts      |  69 ++++++
- .../src/commands/shared/task-mutation.ts           |  60 +++++-
+ .../shared/task-backend-branch-snapshot.ts         | 123 +++++-
+ .../task-backend-branch-snapshot.unit.test.ts      | 136 +++++-
+ .../src/commands/shared/task-backend.test.ts       |  33 +-
+ .../agentplane/src/commands/shared/task-backend.ts |  55 ++-
+ .../src/commands/shared/task-mutation.test.ts      |  69 +++
+ .../src/commands/shared/task-mutation.ts           |  60 ++-
  .../commands/shared/workflow-step-branch-state.ts  |  18 +-
- .../src/commands/shared/workflow-step-branch.ts    |  28 +--
- .../shared/workflow-step-fingerprint.test.ts       |  50 ++++-
- .../commands/shared/workflow-step-fingerprint.ts   |  33 ++-
+ .../src/commands/shared/workflow-step-branch.ts    |  28 +-
+ .../shared/workflow-step-fingerprint.test.ts       |  50 ++-
+ .../commands/shared/workflow-step-fingerprint.ts   |  33 +-
  .../shared/workflow-step-hosted-close.test.ts      |   9 +-
- .../workflow-step-projections-routing.test.ts      |  31 +++
+ .../workflow-step-projections-routing.test.ts      |  31 ++
  .../commands/shared/workflow-step-quality.test.ts  |   6 +-
  .../external-agent-implementation-authority.ts     |   4 +-
- .../agentplane/src/commands/task/finish-shared.ts  | 104 +++++----
- .../src/commands/task/finish-task-store.testkit.ts |  33 +++
- .../commands/task/finish.close-tail.unit.test.ts   |  26 +--
- .../src/commands/task/finish.state.unit.test.ts    |  27 ++-
- .../commands/task/finish.validation.unit.test.ts   |  39 +---
+ .../agentplane/src/commands/task/finish-shared.ts  | 104 +++--
+ .../src/commands/task/finish-task-store.testkit.ts |  63 +++
+ .../commands/task/finish.close-tail.unit.test.ts   |  58 +--
+ .../src/commands/task/finish.state.unit.test.ts    |  27 +-
+ .../commands/task/finish.validation.unit.test.ts   |  39 +-
  .../src/commands/task/hosted-close.command.ts      |   1 +
- .../agentplane/src/commands/task/plan.unit.test.ts |  30 ++-
- .../src/commands/task/set-status.unit.test.ts      |  64 ++++++
- .../task/task-centric-external-result.test.ts      |  85 ++++++++
- .../commands/task/task-centric-external-result.ts  |  56 +++--
- .../task/verify-record.durability.unit.test.ts     |  38 +++-
- 30 files changed, 1291 insertions(+), 287 deletions(-)
+ .../agentplane/src/commands/task/plan.unit.test.ts |  30 +-
+ .../src/commands/task/set-status.unit.test.ts      |  64 +++
+ .../task/task-centric-external-result.test.ts      |  85 ++++
+ .../commands/task/task-centric-external-result.ts  |  56 ++-
+ .../task/verify-record.durability.unit.test.ts     |  38 +-
+ 32 files changed, 1779 insertions(+), 708 deletions(-)
 ```
 
 </details>
